Fixed event registration, this should be documented for future.
