+++ logic.py
@@ -34,21 +34,6 @@ def copy_article_for_update(article_id):
     return article
 
 
-def register_update_time(**kwargs):
-    """
-    Once article published, update its entry in version table to reflect publication date
-    Receives two kwargs: request and article
-    """
-
-    article = kwargs.get('article')
-
-    if hasattr(article, 'version'):
-        v = article.version
-        v.revision_date = article.date_published
-        v.is_published = True
-        v.save()
-
-
 def run_archive(request):
     """
     Function to automatically run an archive containing all published articles
